app/requirements.txt
fastapi
uvicorn[standard]
xgboost
pandas
python-dotenv
joblib
requests